move processing steps to separate script
